requirements.txt
rpi-ws281x
spidev
gpiozero
RPi.GPIO
python-chess
pygame
